Refactor experience events into separate packages.
Level events are now in one package, and XP events are now in a second
package.
