+++ src/main/java/com/gmail/nossr50/events/experience/levels/McMMOPlayerLevelChangeEvent.java
@@ -1,8 +1,9 @@
-package com.gmail.nossr50.events.experience;
+package com.gmail.nossr50.events.experience.levels;
 
 import org.bukkit.entity.Player;
 
 import com.gmail.nossr50.datatypes.skills.SkillType;
+import com.gmail.nossr50.events.experience.McMMOPlayerExperienceEvent;
 
 /**
  * Called when a user levels change
