+++ src/main/java/com/gmail/nossr50/util/EventUtils.java
@@ -15,12 +15,12 @@ import com.gmail.nossr50.datatypes.player.McMMOPlayer;
 import com.gmail.nossr50.datatypes.player.PlayerProfile;
 import com.gmail.nossr50.datatypes.skills.AbilityType;
 import com.gmail.nossr50.datatypes.skills.SkillType;
-import com.gmail.nossr50.events.experience.McMMOPlayerLevelChangeEvent;
-import com.gmail.nossr50.events.experience.McMMOPlayerLevelDownEvent;
-import com.gmail.nossr50.events.experience.McMMOPlayerLevelUpEvent;
-import com.gmail.nossr50.events.experience.McMMOPlayerXpChangeEvent;
-import com.gmail.nossr50.events.experience.McMMOPlayerXpGainEvent;
-import com.gmail.nossr50.events.experience.McMMOPlayerXpLossEvent;
+import com.gmail.nossr50.events.experience.levels.McMMOPlayerLevelChangeEvent;
+import com.gmail.nossr50.events.experience.levels.McMMOPlayerLevelDownEvent;
+import com.gmail.nossr50.events.experience.levels.McMMOPlayerLevelUpEvent;
+import com.gmail.nossr50.events.experience.xp.McMMOPlayerXpChangeEvent;
+import com.gmail.nossr50.events.experience.xp.McMMOPlayerXpGainEvent;
+import com.gmail.nossr50.events.experience.xp.McMMOPlayerXpLossEvent;
 import com.gmail.nossr50.events.fake.FakeBlockBreakEvent;
 import com.gmail.nossr50.events.fake.FakeBlockDamageEvent;
 import com.gmail.nossr50.events.fake.FakePlayerAnimationEvent;
